requirements.txt
aiogram==3.18.0

dotenv~=0.9.9
python-dotenv~=1.0.1
pillow~=11.1.0